+++ examples/python/cudaq_qft.py
@@ -13,6 +13,7 @@ my_sim.init()
 # Choose a simulator backend from CUDA Quantum,
 # e.g., custatevec_fp32 (single-precision)
 my_sim.acc = "cudaq:custatevec_fp32"
+my_sim.gpu_device_id = [0]
 
 # Choose how many qubits to simulate
 num_qubits = 10